+++ ehpc-20230701/alibabacloud_ehpc20230701/models.py
@@ -1102,6 +1102,92 @@ class CreateJobRequestTasksTaskSpecResource(TeaModel):
         return self
 
 
+class CreateJobRequestTasksTaskSpecTaskExecutorContainerEnvironmentVars(TeaModel):
+    def __init__(
+        self,
+        name: str = None,
+        value: str = None,
+    ):
+        self.name = name
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.name is not None:
+            result['Name'] = self.name
+        if self.value is not None:
+            result['Value'] = self.value
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+        return self
+
+
+class CreateJobRequestTasksTaskSpecTaskExecutorContainer(TeaModel):
+    def __init__(
+        self,
+        command: List[str] = None,
+        environment_vars: List[CreateJobRequestTasksTaskSpecTaskExecutorContainerEnvironmentVars] = None,
+        image: str = None,
+        working_dir: str = None,
+    ):
+        self.command = command
+        self.environment_vars = environment_vars
+        self.image = image
+        self.working_dir = working_dir
+
+    def validate(self):
+        if self.environment_vars:
+            for k in self.environment_vars:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.command is not None:
+            result['Command'] = self.command
+        result['EnvironmentVars'] = []
+        if self.environment_vars is not None:
+            for k in self.environment_vars:
+                result['EnvironmentVars'].append(k.to_map() if k else None)
+        if self.image is not None:
+            result['Image'] = self.image
+        if self.working_dir is not None:
+            result['WorkingDir'] = self.working_dir
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Command') is not None:
+            self.command = m.get('Command')
+        self.environment_vars = []
+        if m.get('EnvironmentVars') is not None:
+            for k in m.get('EnvironmentVars'):
+                temp_model = CreateJobRequestTasksTaskSpecTaskExecutorContainerEnvironmentVars()
+                self.environment_vars.append(temp_model.from_map(k))
+        if m.get('Image') is not None:
+            self.image = m.get('Image')
+        if m.get('WorkingDir') is not None:
+            self.working_dir = m.get('WorkingDir')
+        return self
+
+
 class CreateJobRequestTasksTaskSpecTaskExecutorVM(TeaModel):
     def __init__(
         self,
@@ -1144,11 +1230,15 @@ class CreateJobRequestTasksTaskSpecTaskExecutorVM(TeaModel):
 class CreateJobRequestTasksTaskSpecTaskExecutor(TeaModel):
     def __init__(
         self,
+        container: CreateJobRequestTasksTaskSpecTaskExecutorContainer = None,
         vm: CreateJobRequestTasksTaskSpecTaskExecutorVM = None,
     ):
+        self.container = container
         self.vm = vm
 
     def validate(self):
+        if self.container:
+            self.container.validate()
         if self.vm:
             self.vm.validate()
 
@@ -1158,15 +1248,59 @@ class CreateJobRequestTasksTaskSpecTaskExecutor(TeaModel):
             return _map
 
         result = dict()
+        if self.container is not None:
+            result['Container'] = self.container.to_map()
         if self.vm is not None:
             result['VM'] = self.vm.to_map()
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('Container') is not None:
+            temp_model = CreateJobRequestTasksTaskSpecTaskExecutorContainer()
+            self.container = temp_model.from_map(m['Container'])
         if m.get('VM') is not None:
             temp_model = CreateJobRequestTasksTaskSpecTaskExecutorVM()
             self.vm = temp_model.from_map(m['VM'])
+        return self
+
+
+class CreateJobRequestTasksTaskSpecVolumeMount(TeaModel):
+    def __init__(
+        self,
+        mount_options: str = None,
+        mount_path: str = None,
+        volume_driver: str = None,
+    ):
+        self.mount_options = mount_options
+        self.mount_path = mount_path
+        self.volume_driver = volume_driver
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.mount_options is not None:
+            result['MountOptions'] = self.mount_options
+        if self.mount_path is not None:
+            result['MountPath'] = self.mount_path
+        if self.volume_driver is not None:
+            result['VolumeDriver'] = self.volume_driver
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('MountOptions') is not None:
+            self.mount_options = m.get('MountOptions')
+        if m.get('MountPath') is not None:
+            self.mount_path = m.get('MountPath')
+        if m.get('VolumeDriver') is not None:
+            self.volume_driver = m.get('VolumeDriver')
         return self
 
 
@@ -1175,15 +1309,21 @@ class CreateJobRequestTasksTaskSpec(TeaModel):
         self,
         resource: CreateJobRequestTasksTaskSpecResource = None,
         task_executor: List[CreateJobRequestTasksTaskSpecTaskExecutor] = None,
+        volume_mount: List[CreateJobRequestTasksTaskSpecVolumeMount] = None,
     ):
         self.resource = resource
         self.task_executor = task_executor
+        self.volume_mount = volume_mount
 
     def validate(self):
         if self.resource:
             self.resource.validate()
         if self.task_executor:
             for k in self.task_executor:
+                if k:
+                    k.validate()
+        if self.volume_mount:
+            for k in self.volume_mount:
                 if k:
                     k.validate()
 
@@ -1199,6 +1339,10 @@ class CreateJobRequestTasksTaskSpec(TeaModel):
         if self.task_executor is not None:
             for k in self.task_executor:
                 result['TaskExecutor'].append(k.to_map() if k else None)
+        result['VolumeMount'] = []
+        if self.volume_mount is not None:
+            for k in self.volume_mount:
+                result['VolumeMount'].append(k.to_map() if k else None)
         return result
 
     def from_map(self, m: dict = None):
@@ -1211,6 +1355,11 @@ class CreateJobRequestTasksTaskSpec(TeaModel):
             for k in m.get('TaskExecutor'):
                 temp_model = CreateJobRequestTasksTaskSpecTaskExecutor()
                 self.task_executor.append(temp_model.from_map(k))
+        self.volume_mount = []
+        if m.get('VolumeMount') is not None:
+            for k in m.get('VolumeMount'):
+                temp_model = CreateJobRequestTasksTaskSpecVolumeMount()
+                self.volume_mount.append(temp_model.from_map(k))
         return self
 
 
@@ -1653,6 +1802,369 @@ class DeleteJobsResponse(TeaModel):
             self.status_code = m.get('statusCode')
         if m.get('body') is not None:
             temp_model = DeleteJobsResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
+class DescribeJobMetricDataRequest(TeaModel):
+    def __init__(
+        self,
+        array_index: List[int] = None,
+        job_id: str = None,
+        metric_name: str = None,
+        task_name: str = None,
+    ):
+        self.array_index = array_index
+        self.job_id = job_id
+        self.metric_name = metric_name
+        self.task_name = task_name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.array_index is not None:
+            result['ArrayIndex'] = self.array_index
+        if self.job_id is not None:
+            result['JobId'] = self.job_id
+        if self.metric_name is not None:
+            result['MetricName'] = self.metric_name
+        if self.task_name is not None:
+            result['TaskName'] = self.task_name
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ArrayIndex') is not None:
+            self.array_index = m.get('ArrayIndex')
+        if m.get('JobId') is not None:
+            self.job_id = m.get('JobId')
+        if m.get('MetricName') is not None:
+            self.metric_name = m.get('MetricName')
+        if m.get('TaskName') is not None:
+            self.task_name = m.get('TaskName')
+        return self
+
+
+class DescribeJobMetricDataShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        array_index_shrink: str = None,
+        job_id: str = None,
+        metric_name: str = None,
+        task_name: str = None,
+    ):
+        self.array_index_shrink = array_index_shrink
+        self.job_id = job_id
+        self.metric_name = metric_name
+        self.task_name = task_name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.array_index_shrink is not None:
+            result['ArrayIndex'] = self.array_index_shrink
+        if self.job_id is not None:
+            result['JobId'] = self.job_id
+        if self.metric_name is not None:
+            result['MetricName'] = self.metric_name
+        if self.task_name is not None:
+            result['TaskName'] = self.task_name
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ArrayIndex') is not None:
+            self.array_index_shrink = m.get('ArrayIndex')
+        if m.get('JobId') is not None:
+            self.job_id = m.get('JobId')
+        if m.get('MetricName') is not None:
+            self.metric_name = m.get('MetricName')
+        if m.get('TaskName') is not None:
+            self.task_name = m.get('TaskName')
+        return self
+
+
+class DescribeJobMetricDataResponseBody(TeaModel):
+    def __init__(
+        self,
+        data_points: str = None,
+        period: int = None,
+        request_id: str = None,
+    ):
+        self.data_points = data_points
+        self.period = period
+        self.request_id = request_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.data_points is not None:
+            result['DataPoints'] = self.data_points
+        if self.period is not None:
+            result['Period'] = self.period
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DataPoints') is not None:
+            self.data_points = m.get('DataPoints')
+        if m.get('Period') is not None:
+            self.period = m.get('Period')
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        return self
+
+
+class DescribeJobMetricDataResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: DescribeJobMetricDataResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = DescribeJobMetricDataResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
+class DescribeJobMetricLastRequest(TeaModel):
+    def __init__(
+        self,
+        array_index: List[int] = None,
+        job_id: str = None,
+        task_name: str = None,
+    ):
+        self.array_index = array_index
+        self.job_id = job_id
+        self.task_name = task_name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.array_index is not None:
+            result['ArrayIndex'] = self.array_index
+        if self.job_id is not None:
+            result['JobId'] = self.job_id
+        if self.task_name is not None:
+            result['TaskName'] = self.task_name
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ArrayIndex') is not None:
+            self.array_index = m.get('ArrayIndex')
+        if m.get('JobId') is not None:
+            self.job_id = m.get('JobId')
+        if m.get('TaskName') is not None:
+            self.task_name = m.get('TaskName')
+        return self
+
+
+class DescribeJobMetricLastShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        array_index_shrink: str = None,
+        job_id: str = None,
+        task_name: str = None,
+    ):
+        self.array_index_shrink = array_index_shrink
+        self.job_id = job_id
+        self.task_name = task_name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.array_index_shrink is not None:
+            result['ArrayIndex'] = self.array_index_shrink
+        if self.job_id is not None:
+            result['JobId'] = self.job_id
+        if self.task_name is not None:
+            result['TaskName'] = self.task_name
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ArrayIndex') is not None:
+            self.array_index_shrink = m.get('ArrayIndex')
+        if m.get('JobId') is not None:
+            self.job_id = m.get('JobId')
+        if m.get('TaskName') is not None:
+            self.task_name = m.get('TaskName')
+        return self
+
+
+class DescribeJobMetricLastResponseBodyMetrics(TeaModel):
+    def __init__(
+        self,
+        array_index: int = None,
+        metric: str = None,
+    ):
+        self.array_index = array_index
+        self.metric = metric
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.array_index is not None:
+            result['ArrayIndex'] = self.array_index
+        if self.metric is not None:
+            result['Metric'] = self.metric
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ArrayIndex') is not None:
+            self.array_index = m.get('ArrayIndex')
+        if m.get('Metric') is not None:
+            self.metric = m.get('Metric')
+        return self
+
+
+class DescribeJobMetricLastResponseBody(TeaModel):
+    def __init__(
+        self,
+        metrics: List[DescribeJobMetricLastResponseBodyMetrics] = None,
+        request_id: str = None,
+    ):
+        self.metrics = metrics
+        self.request_id = request_id
+
+    def validate(self):
+        if self.metrics:
+            for k in self.metrics:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['Metrics'] = []
+        if self.metrics is not None:
+            for k in self.metrics:
+                result['Metrics'].append(k.to_map() if k else None)
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.metrics = []
+        if m.get('Metrics') is not None:
+            for k in m.get('Metrics'):
+                temp_model = DescribeJobMetricLastResponseBodyMetrics()
+                self.metrics.append(temp_model.from_map(k))
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        return self
+
+
+class DescribeJobMetricLastResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: DescribeJobMetricLastResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = DescribeJobMetricLastResponseBody()
             self.body = temp_model.from_map(m['body'])
         return self
 
@@ -2845,19 +3357,15 @@ class ListExecutorsResponseBody(TeaModel):
     def __init__(
         self,
         executors: List[ListExecutorsResponseBodyExecutors] = None,
-        job_id: str = None,
         page_number: str = None,
         page_size: str = None,
         request_id: str = None,
-        task_name: str = None,
         total_count: str = None,
     ):
         self.executors = executors
-        self.job_id = job_id
         self.page_number = page_number
         self.page_size = page_size
         self.request_id = request_id
-        self.task_name = task_name
         self.total_count = total_count
 
     def validate(self):
@@ -2876,16 +3384,12 @@ class ListExecutorsResponseBody(TeaModel):
         if self.executors is not None:
             for k in self.executors:
                 result['Executors'].append(k.to_map() if k else None)
-        if self.job_id is not None:
-            result['JobId'] = self.job_id
         if self.page_number is not None:
             result['PageNumber'] = self.page_number
         if self.page_size is not None:
             result['PageSize'] = self.page_size
         if self.request_id is not None:
             result['RequestId'] = self.request_id
-        if self.task_name is not None:
-            result['TaskName'] = self.task_name
         if self.total_count is not None:
             result['TotalCount'] = self.total_count
         return result
@@ -2897,16 +3401,12 @@ class ListExecutorsResponseBody(TeaModel):
             for k in m.get('Executors'):
                 temp_model = ListExecutorsResponseBodyExecutors()
                 self.executors.append(temp_model.from_map(k))
-        if m.get('JobId') is not None:
-            self.job_id = m.get('JobId')
         if m.get('PageNumber') is not None:
             self.page_number = m.get('PageNumber')
         if m.get('PageSize') is not None:
             self.page_size = m.get('PageSize')
         if m.get('RequestId') is not None:
             self.request_id = m.get('RequestId')
-        if m.get('TaskName') is not None:
-            self.task_name = m.get('TaskName')
         if m.get('TotalCount') is not None:
             self.total_count = m.get('TotalCount')
         return self